requirements.txt
PyQt6==6.6.1
python-docx==0.8.11
pywin32==306
pyinstaller==6.3.0
regex==2023.12.25
mysql-connector-python==8.3.0 